tests: Run N-square LB tests in CI
