+++ testsuite/python/CMakeLists.txt
@@ -101,6 +101,7 @@ python_test(FILE dds-and-bh-gpu.py MAX_NUM_PROC 4 LABELS gpu)
 python_test(FILE electrostaticInteractions.py MAX_NUM_PROC 2)
 python_test(FILE engine_langevin.py MAX_NUM_PROC 4)
 python_test(FILE engine_lb.py MAX_NUM_PROC 2 LABELS gpu)
+python_test(FILE engine_lb_nsquare.py MAX_NUM_PROC 1 LABELS gpu)
 python_test(FILE experimental_decorator.py MAX_NUM_PROC 1)
 python_test(FILE icc.py MAX_NUM_PROC 4)
 python_test(FILE mass-and-rinertia_per_particle.py MAX_NUM_PROC 2 LABELS long)
@@ -218,6 +219,7 @@ python_test(FILE lb_shear.py MAX_NUM_PROC 2 LABELS gpu)
 python_test(FILE lb_thermostat.py MAX_NUM_PROC 2 LABELS gpu)
 python_test(FILE lb_buoyancy_force.py MAX_NUM_PROC 4 LABELS gpu)
 python_test(FILE lb_momentum_conservation.py MAX_NUM_PROC 4 LABELS gpu)
+python_test(FILE lb_momentum_conservation_nsquare.py MAX_NUM_PROC 1 LABELS gpu)
 python_test(FILE p3m_electrostatic_pressure.py MAX_NUM_PROC 2)
 python_test(FILE sigint.py DEPENDENCIES sigint_child.py MAX_NUM_PROC 1)
 python_test(FILE lb_density.py MAX_NUM_PROC 1)
