+++ testsuite/python/engine_lb.py
@@ -14,11 +14,12 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+import espressomd
+import espressomd.lb
 import unittest as ut
 import unittest_decorators as utx
 import numpy as np
-import espressomd
-import espressomd.lb
 
 
 class SwimmerTest():
@@ -31,6 +32,7 @@ class SwimmerTest():
                  'kT': 0,
                  'tau': system.time_step}
     gamma = 0.3
+    n_nodes = system.cell_system.get_state()['n_nodes']
 
     def add_all_types_of_swimmers(
             self,
@@ -69,14 +71,7 @@ class SwimmerTest():
                                   "dipole_length": 0.8})
 
     def setUp(self):
-        if self.decomposition_type == "regular":
-            self.system.cell_system.set_regular_decomposition()
-        elif self.decomposition_type == "n_square":
-            if any(self.system.cell_system.node_grid > 1):
-                self.skipTest(
-                    "N Square and LB not compatible on more than one core")
-            self.system.cell_system.set_n_square()
-
+        self.set_cellsystem()
         self.lbf = self.lb_class(**self.LB_params)
         self.system.actors.add(self.lbf)
         self.system.thermostat.set_lb(LB_fluid=self.lbf, gamma=self.gamma)
@@ -136,35 +131,22 @@ class SwimmerTest():
 @utx.skipIfMissingFeatures(["ENGINE", "ROTATIONAL_INERTIA", "MASS"])
 class SwimmerTestRegularCPU(SwimmerTest, ut.TestCase):
 
-    decomposition_type = "regular"
     lb_class = espressomd.lb.LBFluid
     tol = 1e-10
 
-
-@utx.skipIfMissingFeatures(["ENGINE", "ROTATIONAL_INERTIA", "MASS"])
-class SwimmerTestNSquareCPU(SwimmerTest, ut.TestCase):
-
-    decomposition_type = "n_square"
-    lb_class = espressomd.lb.LBFluid
-    tol = 1e-10
+    def set_cellsystem(self):
+        self.system.cell_system.set_regular_decomposition()
 
 
 @utx.skipIfMissingGPU()
 @utx.skipIfMissingFeatures(["ENGINE", "ROTATIONAL_INERTIA", "MASS"])
 class SwimmerTestRegularGPU(SwimmerTest, ut.TestCase):
 
-    decomposition_type = "regular"
     lb_class = espressomd.lb.LBFluidGPU
     tol = 1e-5
 
-
-@utx.skipIfMissingGPU()
-@utx.skipIfMissingFeatures(["ENGINE", "ROTATIONAL_INERTIA", "MASS"])
-class SwimmerTestNSquareGPU(SwimmerTest, ut.TestCase):
-
-    decomposition_type = "n_square"
-    lb_class = espressomd.lb.LBFluidGPU
-    tol = 1e-5
+    def set_cellsystem(self):
+        self.system.cell_system.set_regular_decomposition()
 
 
 if __name__ == "__main__":
